src/backend/styletransfer/requirements.txt
Pillow==10.1.*
python-dotenv==1.0.*
redis==5.0.*
torch==2.1.*
torchvision==0.16.*